(refactor) : WinningInformation 수정 - 2등 추가 - 보너스 볼 포함 여부 추가 - findByCountOfMatchedNumber 수정
